pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "skillvector-engine"
version = "0.2.0"
description = "AI-powered career skill gap analysis engine"
requires-python = ">=3.11"
dependencies = [
    "langchain-anthropic>=0.1.0,<2.0.0",
    "langchain-core>=0.2.0,<2.0.0",
    "sentence-transformers>=2.2.0,<3.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "numpy>=1.24.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pinecone-client>=3.0.0,<5.0.0",
    "neo4j>=5.0.0,<6.0.0",
    "streamlit>=1.30.0,<2.0.0",
    "pdfplumber>=0.10.0,<1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<9.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
check_untyped_defs = true
